build: use bignum-kt 1.0.4's signum constructor instead of a local helper
1.0.4 adds `BigInteger(signum, magnitude)` and its offset/length overload, so
the eight unsigned-magnitude sites go back to `BigInteger(1, bytes)` and the
four duplicated `bigIntegerFromUnsigned` helpers are deleted. Net -101 lines.

The helper only existed because those four modules had nowhere to share code:
ethers-rlp and ethers-crypto are both leaves. Upstreaming it was the right fix.

1.0.4 still has no companion object, so `bigIntegerOf(..)` remains the way to
get ZERO/ONE/TWO and is unaffected.

The test is retargeted rather than dropped. It now pins the library's
behaviour instead of ours, which is still worth having: the single-argument
constructor reads two's-complement, so silently swapping it for the signum form
corrupts values rather than failing, and this catches that on a future upgrade.

Re-verified with a throwaway iosSimulatorArm64 target: ethers-rlp still
compiles for iOS with zero errors.
